LayoutTests/fast/dom/Document/compatMode-location-expected.txt

Tests that the compatMode property getter is on Document and not HTMLDocument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.getOwnPropertyDescriptor(HTMLDocument.prototype, 'compatMode') is undefined.
PASS Object.getOwnPropertyDescriptor(Document.prototype, 'compatMode').get is an instance of Function
PASS successfullyParsed is true

TEST COMPLETE
